Move ssl engine related util methods to a class.
Motivation:

A couple of static utils method to deal with SSL engine have been added.

We can consolidate them in an SSL engine related utils method.
