--- vertx-core/src/main/java/io/vertx/core/internal/net/SslEngineHelper.java
@@ -1,17 +1,0 @@
-/*
- * Copyright (c) 2011-2026 Contributors to the Eclipse Foundation
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
- * which is available at https://www.apache.org/licenses/LICENSE-2.0.
- *
- * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
- */
-package io.vertx.core.internal.net;
-
-/**
- * Pre Java 21 implementation: named groups not supported.
- */
-class SslEngineHelper {
-}
